+++ tests/TestCase.php
@@ -1,5 +1,7 @@
 <?php
 
+use App\Http\Controllers\Controller;
+
 class TestCase extends Illuminate\Foundation\Testing\TestCase
 {
     /**
@@ -21,5 +23,25 @@ class TestCase extends Illuminate\Foundation\Testing\TestCase
         $app->make(Illuminate\Contracts\Console\Kernel::class)->bootstrap();
 
         return $app;
+    }
+
+    /**
+     * Calls a protected method on an abstract class
+     *
+     * @author EB
+     * @param $class
+     * @param $method
+     * @param $params
+     * @return mixed
+     */
+    protected function callProtectedMethodOnAbstractClass($method, $params, $class = Controller::class)
+    {
+        $reflection = new ReflectionClass($class);
+        $method = $reflection->getMethod($method);
+        $method->setAccessible(true);
+
+        $obj = $this->getMockForAbstractClass($class);
+
+        return $method->invokeArgs($obj, $params);
     }
 }